+++ services/operator-value-parser.js
@@ -34,6 +34,7 @@ var VALUES_DATE = [
 
 var PERIODS_PREVIOUS_X_DAYS = /^\$previous(\d+)Days$/;
 var PERIODS_X_DAYS_TO_DATE = /^\$(\d+)DaysToDate$/;
+var PERIODS_X_HOURS_BEFORE = /^\$(\d+)HoursBefore$/;
 
 function OperatorValueParser(opts, timezone) {
 
@@ -92,7 +93,11 @@ function OperatorValueParser(opts, timezone) {
     function isIntervalDateValue(value) {
       var match = value.match(PERIODS_PREVIOUS_X_DAYS);
       if (match && match[1]) { return true; }
+
       match = value.match(PERIODS_X_DAYS_TO_DATE);
+      if (match && match[1]) { return true; }
+
+      match = value.match(PERIODS_X_HOURS_BEFORE);
       if (match && match[1]) { return true; }
 
       return VALUES_DATE.indexOf(value) !== -1;
@@ -136,6 +141,11 @@ function OperatorValueParser(opts, timezone) {
                   .startOf('day').toDate(),
           $lte: moment().toDate()
         };
+      }
+
+      match = value.match(PERIODS_X_HOURS_BEFORE);
+      if (match && match[1]) {
+        return { $lte: moment().subtract(match[1], 'days').toDate() };
       }
 
       switch (value) {